+++ Pages/ECC/ecc.html
@@ -34,7 +34,7 @@
     <form action="../index.html">
         <button type="submit">Back to index</button>
     </form>
-    <script src="scripts/helpers.mjs"></script>
+    <script src="scripts/eccHelpers.mjs"></script>
     <script src="scripts/maps_stuff.mjs" Cache-Control: public></script>
     <script src="scripts/ecc.js"></script>
 </body>
