+++ Pages/index.html
@@ -3,7 +3,7 @@
 
 <head>
     <link href="https://fonts.googleapis.com/css?family=Inter&display=swap" rel="stylesheet" />
-    <link href="/pages/main.css" rel="stylesheet"/>
+    <link href="/pages/main.css" rel="stylesheet" />
     <title>Emergency Helpdesk</title>
 </head>
 
@@ -12,8 +12,12 @@
         <div class="v2_6"></div>
         <div class="v2_16">
             <div class="v2_7"></div><span class="v2_8">Emergency Helpdesk</span>
-            <a href="/Pages/Caller/caller.html"><div class="v2_9"></div><span class="v2_17">Caller</span></a>
-            <a href="/Pages/ECC/ecc.html"><div class="v2_10"></div><span class="v2_18">ECC Operator</span></a>
+            <a href="/Pages/Caller/caller.html">
+                <div class="v2_9"></div><span class="v2_17">Caller</span>
+            </a>
+            <a href="/Pages/ECC/ecc.html">
+                <div class="v2_10"></div><span class="v2_18">ECC Operator</span>
+            </a>
         </div>
     </div>
 </body>
