+++ Server/FunctionFiles/callerobj.mjs
@@ -1,8 +1,8 @@
 import * as fs from 'fs'
 
-import { exportObject } from '../helpers.mjs';
-import { getPostData } from "../helpers.mjs";
-import { addCaller } from '../helpers.mjs';
+import { exportObject } from '../serverHelpers.mjs';
+import { getPostData } from "../serverHelpers.mjs";
+import { addCaller } from '../serverHelpers.mjs';
 import { errorResponse } from '../responseHandlers.mjs';
 
 export function pageCallerObj(req, res, path) {
