+++ Server/FunctionFiles/change_answering.mjs
@@ -1,4 +1,4 @@
-import { getPostData, importObject, exportObject } from "../helpers.mjs"
+import { getPostData, importObject, exportObject } from "../serverHelpers.mjs"
 import { errorResponse } from "../responseHandlers.mjs";
 
 export function pageChangeAnswering(req, res, path) {
