+++ Server/responseHandlers.mjs
@@ -3,8 +3,8 @@ import { pageEmergencyAccepted } from "./FunctionFiles/emergency_accepted.mjs";
 import { pageEmergencyHandled } from "./FunctionFiles/emergency_handled.mjs";
 import { pageEcc as pageEcc } from "./FunctionFiles/ecc.mjs";
 import { pageChangeAnswering as pageChangeAnswering } from "./FunctionFiles/change_answering.mjs"
-import { determineMimeType, addCaller } from "./helpers.mjs";
-import { getArgs as getArgs } from "./helpers.mjs";
+import { determineMimeType, addCaller } from "./serverHelpers.mjs";
+import { getArgs as getArgs } from "./serverHelpers.mjs";
 import * as fs from 'fs';
 import { operator } from "./classes.mjs";
